Revert that PR #1017 added new HeroBackgroundImage and center image with flexbox instead
This works both for LandingPage and OfferPage and seems simpler than
adjusting object-size with breakpoints
