+++ src/client/components/BackgroundImage.tsx
@@ -1,29 +1,28 @@
 import React, { useState } from 'react'
 import styled from '@emotion/styled'
-import { css } from '@emotion/core'
 import { colorsV3 } from '@hedviginsurance/brand'
-import { LARGE_SCREEN_MEDIA_QUERY } from 'utils/mediaQueries'
+
+type BackgroundImageVariant = 'home' | 'car'
 
 type Props = {
-  height?: number
   isFullScreen?: boolean
   zIndex: number
   className?: string
+  variant?: BackgroundImageVariant
 }
 
 const ImageContainer = styled.div<Props>`
   background-color: ${colorsV3.gray900};
   width: 100%;
-  height: ${({ height }) => (height ? `${height}px` : '100%')};
+  height: ${({ isFullScreen }) => (isFullScreen ? '100vh' : '100%')};
   position: absolute;
   top: 0;
   overflow: hidden;
   z-index: ${({ zIndex }) => zIndex};
-  ${({ isFullScreen }) =>
-    isFullScreen &&
-    css`
-      height: 100vh;
-    `}
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
 `
 
 type ImageProps = {
@@ -39,23 +38,27 @@ const Image = styled.img<ImageProps>`
   opacity: ${({ hasLoaded }) => (hasLoaded ? 0.6 : 0)};
   transition: opacity 1s cubic-bezier(0.33, 1, 0.68, 1);
   transition-delay: 150ms;
-  object-position: ${({ isFullScreen }) =>
-    isFullScreen ? '50% 50%' : '100% -100px'};
-
-  ${LARGE_SCREEN_MEDIA_QUERY} {
-    object-position: 100% 50%;
-  }
 `
 
-export const BackgroundImage = ({
-  height,
-  isFullScreen = false,
-  zIndex,
-  className,
-}: Props) => {
-  const [hasImageLoaded, setHasImageLoaded] = useState(false)
+const PictureElement = styled.picture<{ isFullScreen?: boolean }>(
+  ({ isFullScreen }) => ({
+    width: '100%',
+    ...(isFullScreen && { height: '100%' }),
+  }),
+)
 
-  const images = {
+type ImageVariantSizes = {
+  small: string
+  medium: string
+}
+
+type ImageVariant = {
+  portrait: ImageVariantSizes
+  landscape: ImageVariantSizes
+}
+
+const variants: Record<BackgroundImageVariant, ImageVariant> = {
+  home: {
     portrait: {
       small: '/new-member-assets/landing/hedvig_living_room_portrait_small.jpg',
       medium:
@@ -67,15 +70,36 @@ export const BackgroundImage = ({
       medium:
         '/new-member-assets/landing/hedvig_living_room_landscape_medium.jpg',
     },
-  }
+  },
+  car: {
+    portrait: {
+      small: '/new-member-assets/landing/hedvig_garage_portrait_small.jpg',
+      medium: '/new-member-assets/landing/hedvig_garage_portrait_medium.jpg',
+    },
+    landscape: {
+      small: '/new-member-assets/landing/hedvig_garage_landscape_small.jpg',
+      medium: '/new-member-assets/landing/hedvig_garage_landscape_medium.jpg',
+    },
+  },
+}
+
+export const BackgroundImage = ({
+  className,
+  isFullScreen = false,
+  zIndex,
+  variant = 'home',
+}: Props) => {
+  const [hasImageLoaded, setHasImageLoaded] = useState(false)
+
+  const images = variants[variant]
+
   return (
     <ImageContainer
       className={className}
-      height={height}
       isFullScreen={isFullScreen}
       zIndex={zIndex}
     >
-      <picture>
+      <PictureElement isFullScreen={isFullScreen}>
         <source
           media="(orientation: portrait)"
           srcSet={`${images.portrait.small} 900w, ${images.portrait.medium} 1600w`}
@@ -90,7 +114,7 @@ export const BackgroundImage = ({
           hasLoaded={hasImageLoaded}
           isFullScreen={isFullScreen}
         />
-      </picture>
+      </PictureElement>
     </ImageContainer>
   )
 }
